# api-docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "3.0.2",
    "title": "CVE Services API",
    "description": "The CVE services API which supports automation tooling for the CVE Project.",
    "contact": {
      "name": "CVE Services",
      "url": "https://cveawg-dev.mitre.org/api",
      "email": "cve-board-auto-list@mitre.org"
    }
  },
  "servers": [
    {
      "url": "https://cveawg-dev.mitre.org/api"
    }
  ],
  "paths": {
...
    "/cve": {
      "get": {
        "tags": [
          "CVE Record"
        ],
        "summary": "Retrieves all CVE records (This endpoint can only be used by the Secretariat.)",
        "description": "  This endpoint is an internal admin endpoint restricted to the Secretariat.  <h2>Access Control</h2>  <p>At least one of the following roles are needed to access the endpoint:</p>  <p>- <b>SECRETARIAT:</b> The user must belong to an Organization with the “SECRETARIAT” role</p>  <h2>Expected Behavior</h2>  <p><b>Secretariat:</b> Can see all CVE records owned by any Organization</p>",
        "operationId": "cveGetFiltered",
        "parameters": [
          {
            "$ref": "#/components/parameters/cveRecordFilteredTimeModifiedLt"
          },
          {
            "$ref": "#/components/parameters/cveRecordFilteredTimeModifiedGt"
          },
          {
            "$ref": "#/components/parameters/cveState"
          },
          {
            "$ref": "#/components/parameters/countOnly"
          },
          {
            "$ref": "#/components/parameters/assignerShortName"
          },
          {
            "$ref": "#/components/parameters/assigner"
          },
          {
            "$ref": "#/components/parameters/apiEntityHeader"
          },
          {
            "$ref": "#/components/parameters/apiUserHeader"
          },
          {
            "$ref": "#/components/parameters/apiSecretHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of CVE records and pagination fields, only allowed by the Secretariat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/cve/list-cve-records-response.json"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/bad-request.json"
                }
              }
            }
          },
          "401": {
            "description": "Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          }
        }
      }
    },
    "/cve/{id}/cna": {
      "post": {
        "tags": [
          "CVE Record"
        ],
        "summary": "Creates a CVE Record by ID (This endpoint is to be used by CNAs.)",
        "description": "  <h2>Access Control</h2>  <p>No roles needed to access the endpoint</p>  <h2>Expected Behavior</h2>  <p><b>Admin User:</b> Can submit CVE records for CVE IDs that are owned by the organization</p>  <p><b>Regular User:</b> Can submit CVE records for CVE IDs that are owned by the organization</p>",
        "operationId": "cveCnaCreateSingle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The CVE ID for which the record is being created."
          },
          {
            "$ref": "#/components/parameters/apiEntityHeader"
          },
          {
            "$ref": "#/components/parameters/apiUserHeader"
          },
          {
            "$ref": "#/components/parameters/apiSecretHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested CVE record is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/cve/create-cve-record-response.json"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/bad-request.json"
                }
              }
            }
          },
          "401": {
            "description": "Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "/schemas/cve/cve-record-minimum-request.json"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CVE Record"
        ],
        "summary": "Update a CVE Record by ID (This endpoint is to be used by CNAs.)",
        "description": "  <h2>Access Control</h2>  <p>No roles needed to access the endpoint</p>  <h2>Expected Behavior</h2>  <p><b>Admin User:</b> Can submit CVE records for CVE IDs that are owned by the organization</p>  <p><b>Regular User:</b> Can submit CVE records for CVE IDs that are owned by the organization</p>",
        "operationId": "cveCnaUpdateSingle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The CVE ID for which the record is being updated."
          },
          {
            "$ref": "#/components/parameters/apiEntityHeader"
          },
          {
            "$ref": "#/components/parameters/apiUserHeader"
          },
          {
            "$ref": "#/components/parameters/apiSecretHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the CVE record updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/cve/update-cve-record-response.json"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/bad-request.json"
                }
              }
            }
          },
          "401": {
            "description": "Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/schemas/errors/generic.json"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "/schemas/cve/cve-record-minimum-request.json"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "apiEntityHeader": {
        "in": "header",
        "name": "CVE-API-ORG",
        "description": "The shortname for the entity (e.g., CNA, ADP) that is authenticated or requesting authentication.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "apiUserHeader": {
        "in": "header",
        "name": "CVE-API-USER",
        "description": "The username for the account that is making the request.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "apiSecretHeader": {
        "in": "header",
        "name": "CVE-API-KEY",
        "description": "The user's API key.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "assigner": {
        "in": "query",
        "name": "assigner",
        "description": "Specify the UUID for the org to which the cve id was assigned.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "assignerShortName": {
        "in": "query",
        "name": "assignerShortName",
        "description": "Get the short name for the org to which the cve id was assigned.",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
      "countOnly": {
        "in": "query",
        "name": "count_only",
        "description": "Get count of records that match query. Accepted values are 0 and 1, representing False and True, respectively.",
        "required": false,
        "schema": {
          "type": "boolean"
        }
      },
      "cveState": {
        "in": "query",
        "name": "state",
        "description": "Filter by CVE state.",
        "schema": {
          "type": "string",
          "enum": [
            "PUBLISHED",
            "REJECTED"
          ]
        }
      },
...
      "cveRecordFilteredTimeModifiedLt": {
        "in": "query",
        "name": "time_modified.lt",
        "description": "Get all CVE records modified before timestamp.",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "cveRecordFilteredTimeModifiedGt": {
        "in": "query",
        "name": "time_modified.gt",
        "description": "Get all CVE records modified after timestamp.",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
...
    }
  }
}